requirements.txt
fastapi
uvicorn
pandas
groq
python-dotenv
pydantic
